iths_python/projekt/requirements.txt
cryptography
requests
python-nmap
scapy